service/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
python-multipart==0.0.20
PyYAML==6.0.2
scikit-learn==1.6.0
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
uvicorn==0.34.0
